+++ src/Thoth.Json.Net/RELEASE_NOTES.md
@@ -1,6 +1,10 @@
+#### 2.4.0
+
+* Make auto decoder succeeds on Class marked as optional
+
 #### 2.3.0
 
-* Fix #59: Make autodecoder support optional fields when missing from JSON
+* Fix #59: Make auto decoder support optional fields when missing from JSON
 * Fix #61: Support object keys with JsonPath characters when using `Decode.dict`
 * Fix #51: Add support for `Raw` decoder in object builders
 * Added CultureInfo.InvariantCulture to all Encoder functions where it was possible (by @draganjovanovic1)